feat: implement updatePlatformVideo function and remove updateTutorial function
